add convergenza serie

+++ fibonacci.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int n, t1 = 0, t2 = 1, nextTerm = 0;
+    long long n, t1 = 0, t2 = 1, nextTerm = 0;
     cout << "Enter the number of terms: ";
     cin >> n;
